Reorganize Code + Fix other InvestmentModel Graph
